+++ CHANGELOG.md
@@ -2,11 +2,17 @@
 
 ### New Features
 
-* Added support for [graph-gophers/graphql-go](https://github.com/graph-gophers/graphql-go)
-  applications in the new [v3/integrations/nr-graph-gophers](https://godoc.org/github.com/newrelic/go-agent/v3/integrations/nr-graph-gophers)
-  package.
-  * [Documentation](https://godoc.org/github.com/newrelic/go-agent/v3/integrations/nr-graph-gophers)
-  * [Example](v3/integrations/nr-graph-gophers/example/main.go)
+* Added support for GraphQL in two new integrations.
+  * [graph-gophers/graphql-go](https://github.com/graph-gophers/graphql-go)
+  with
+  [v3/integrations/nr-graph-gophers](https://godoc.org/github.com/newrelic/go-agent/v3/integrations/nr-graph-gophers).
+    * [Documentation](https://godoc.org/github.com/newrelic/go-agent/v3/integrations/nr-graph-gophers)
+    * [Example](v3/integrations/nr-graph-gophers/example/main.go)
+  * [graphql-go/graphql](https://github.com/graphql-go/graphql)
+  with
+  [v3/integrations/nr-graphql-go](https://godoc.org/github.com/newrelic/go-agent/v3/integrations/nr-graphql-go).
+    * [Documentation](https://godoc.org/github.com/newrelic/go-agent/v3/integrations/nr-graphql-go)
+    * [Example](v3/integrations/nr-graphql-go/example/main.go)
 
 * Added database instrumentation support for
   [snowflakedb/gosnowflake](https://github.com/snowflakedb/gosnowflake).
@@ -21,6 +27,12 @@
   [`newrelic.NewRoundTripper`](https://godoc.org/github.com/newrelic/go-agent/v3/newrelic#NewRoundTripper),
   if existing cross application tracing or distributed tracing headers are
   present on the request, they will be replaced instead of added.
+
+* The
+  [`FromContext`](https://godoc.org/github.com/newrelic/go-agent/v3/newrelic#FromContext)
+  API which allows you to pull a Transaction from a context.Context will no
+  longer panic if the provided context is nil.  In this case, a nil is
+  returned.
 
 ## 3.2.0